split package check fixes
